replace int with long

+++ src/main/java/abstractdomain/Domain.java
@@ -3,9 +3,9 @@ package abstractdomain;
 import abstractdomain.exception.ConcretizationException;
 
 public interface Domain<T extends DomainValue<T>> {
-  T abstract_(int concreteValue);
+  T abstract_(long concreteValue);
 
-  int concretize(T abstractValue) throws ConcretizationException;
+  long concretize(T abstractValue) throws ConcretizationException;
 
   T getUnknown();
